Revert "DEP-6484: Conversion of controllers folder to TS"
