+++ app/assets/javascripts/controllers/ReactiveChatController.js
@@ -3,16 +3,8 @@ import ClickToChatButton from '../utils/ClickToChatButton'
 import CommonChatController from './CommonChatController'
 import * as DisplayState from '../NuanceDisplayState'
 import { messages} from '../utils/Messages'
-import { ClickToChatObjectInterface } from '../utils/ClickToChatButtons'
 
-interface c2cDisplayStateMessagesInterface {
-    [DisplayState.OutOfHours]: string
-    [DisplayState.Ready]: string
-    [DisplayState.Busy]: string
-    [DisplayState.ChatActive]: string
-}
-
-const c2cDisplayStateMessages: c2cDisplayStateMessagesInterface = {
+const c2cDisplayStateMessages = {
     [DisplayState.OutOfHours]: messages.outofhours,
     [DisplayState.Ready]: messages.ready,
     [DisplayState.Busy]: messages.busy,
@@ -20,20 +12,17 @@ const c2cDisplayStateMessages: c2cDisplayStateMessagesInterface = {
 };
 
 export default class ReactiveChatController {
-    sdk: any
-    c2cButtons: ClickToChatButtons
-    commonChatController: CommonChatController
     constructor() {
         this.sdk = null;
         this.c2cButtons = new ClickToChatButtons(this._clickToChatCallback(), c2cDisplayStateMessages);
         this.commonChatController = new CommonChatController();
     }
 
-    _clickToChatCallback(): (c2cIdx: any) => void {
-        return (c2cIdx: any) => this._onC2CButtonClicked(c2cIdx)
+    _clickToChatCallback() {
+        return (c2cIdx) => this._onC2CButtonClicked(c2cIdx)
     }
 
-    addC2CButton(c2cObj: ClickToChatObjectInterface, divID: string, buttonClass: string): void {
+    addC2CButton(c2cObj, divID, buttonClass) {
         if (c2cObj.displayState == "ready") {
             this.c2cButtons.addButton(
                 c2cObj,
@@ -43,12 +32,12 @@ export default class ReactiveChatController {
         }
     }
 
-    _onC2CButtonClicked(c2cIdx: any): void {
-        const reactiveObj: {type: string} = {
+    _onC2CButtonClicked(c2cIdx) {
+        const reactiveObj= {
             type: 'reactive'
         }
         this.sdk = window.Inq.SDK;
-        this.sdk.onC2CClicked(c2cIdx, () => {
+        this.sdk.onC2CClicked(c2cIdx, (state) => {
             this.commonChatController._launchChat(reactiveObj);
         });
     }
