+++ app/assets/javascripts/index.d.ts
@@ -3,7 +3,5 @@ export {};
 declare global {
   interface Window {
     chatId: number
-    featureSwitchUrl: any
-    Inq: any
   }
 }
